requirements.txt
altair==5.5.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
cachetools==6.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.3
debugpy==1.8.15
decorator==5.2.1
dotenv==0.9.9
executing==2.2.0
flash==1.0.3
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
ipykernel==6.30.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
narwhals==2.0.1
nest-asyncio==1.6.0
numpy==2.3.2
packaging==25.0
pandas==2.3.1
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
polars==1.31.0
prettyprint==0.1.5
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pydeck==0.9.1
Pygments==2.19.2
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
six==1.17.0
smmap==5.0.2
soupsieve==2.7
stack-data==0.6.3
streamlit==1.47.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
